requirements.txt
fastapi
uvicorn[standard]
pyodbc
python-dotenv

